+++ tns-core-modules/ui/core/view/view.android.ts
@@ -1,7 +1,6 @@
 // Definitions.
 import { Point, CustomLayoutView as CustomLayoutViewDefinition, dip } from ".";
 import { GestureTypes, GestureEventData } from "../../gestures";
-import { AndroidActivityBackPressedEventData } from "../../../application";
 // Types.
 import {
     ViewCommon, layout, isEnabledProperty, originXProperty, originYProperty, automationTextProperty, isUserInteractionEnabledProperty,
@@ -22,6 +21,7 @@ import {
 import { Background, ad as androidBackground } from "../../styling/background";
 import { profile } from "../../../profiling";
 import { topmost } from "../../frame/frame-stack";
+import { AndroidActivityBackPressedEventData, android as androidApp } from "../../../application";
 
 export * from "./view-common";
 
@@ -72,12 +72,7 @@ function initializeTouchListener(): void {
 
         onTouch(view: android.view.View, event: android.view.MotionEvent): boolean {
             const owner = this.owner;
-            for (let type in owner._gestureObservers) {
-                let list = owner._gestureObservers[type];
-                list.forEach(element => {
-                    element.androidOnTouchEvent(event);
-                });
-            }
+            owner.handleGestureTouch(event);
 
             let nativeView = owner.nativeViewProtected;
             if (!nativeView || !nativeView.onTouchEvent) {
@@ -112,6 +107,13 @@ function initializeDialogFragment() {
                 activity: view._context,
                 cancel: false,
             };
+
+            // Fist fire application.android global event
+            androidApp.notify(args);
+            if (args.cancel) {
+                return;
+            }
+
             view.notify(args);
 
             if (!args.cancel && !view.onBackPressed()) {
@@ -178,7 +180,7 @@ function initializeDialogFragment() {
             }
 
             const owner = this.owner;
-            if (!owner.isLoaded) {
+            if (owner && !owner.isLoaded) {
                 owner.callLoaded();
             }
 
@@ -194,7 +196,7 @@ function initializeDialogFragment() {
             }
 
             const owner = this.owner;
-            if (owner.isLoaded) {
+            if (owner && owner.isLoaded) {
                 owner.callUnloaded();
             }
         }
@@ -232,6 +234,7 @@ export class View extends ViewCommon {
     private layoutChangeListenerIsSet: boolean;
     private layoutChangeListener: android.view.View.OnLayoutChangeListener;
     private _manager: android.support.v4.app.FragmentManager;
+    private _rootManager: android.support.v4.app.FragmentManager;
 
     nativeViewProtected: android.view.View;
 
@@ -263,24 +266,56 @@ export class View extends ViewCommon {
         }
     }
 
+    public _getChildFragmentManager(): android.support.v4.app.FragmentManager {
+        return null;
+    }
+
+    public _getRootFragmentManager(): android.support.v4.app.FragmentManager {
+        if (!this._rootManager && this._context) {
+            this._rootManager = (<android.support.v4.app.FragmentActivity>this._context).getSupportFragmentManager();
+        }
+
+        return this._rootManager;
+    }
+
     public _getFragmentManager(): android.support.v4.app.FragmentManager {
         let manager = this._manager;
         if (!manager) {
             let view: View = this;
+            let frameOrTabViewItemFound = false;
             while (view) {
+                // when interacting with nested fragments instead of using getSupportFragmentManager 
+                // we must always use getChildFragmentManager instead;
+                // we have three sources of fragments -- Frame fragments, TabViewItem fragments, and
+                // modal dialog fragments
+
+                // modal -> frame / tabview (frame / tabview use modal CHILD fm)
                 const dialogFragment = view._dialogFragment;
                 if (dialogFragment) {
                     manager = dialogFragment.getChildFragmentManager();
                     break;
-                } else {
-                    // the case is needed because _dialogFragment is on View
-                    // but parent may be ViewBase.
-                    view = view.parent as View;
                 }
+
+                // - frame1 -> frame2 (frame2 uses frame1 CHILD fm)
+                // - tabview -> frame1 (frame1 uses tabview item CHILD fm)
+                // - frame1 -> tabview (tabview uses frame1 CHILD fm)
+                // - frame1 -> tabview -> frame2 (tabview uses frame1 CHILD fm; frame2 uses tabview item CHILD fm)
+                if (view._hasFragments) {
+                    if (frameOrTabViewItemFound) {
+                        manager = view._getChildFragmentManager();
+                        break;
+                    }
+
+                    frameOrTabViewItemFound = true;
+                }
+
+                // the case is needed because _dialogFragment is on View
+                // but parent may be ViewBase.
+                view = view.parent as View;
             }
 
-            if (!manager && this._context) {
-                manager = (<android.support.v4.app.FragmentActivity>this._context).getSupportFragmentManager();
+            if (!manager) {
+                manager = this._getRootFragmentManager();
             }
 
             this._manager = manager;
@@ -292,6 +327,7 @@ export class View extends ViewCommon {
     @profile
     public onLoaded() {
         this._manager = null;
+        this._rootManager = null;
         super.onLoaded();
         this.setOnTouchListener();
     }
@@ -305,6 +341,7 @@ export class View extends ViewCommon {
         }
 
         this._manager = null;
+        this._rootManager = null;
         super.onUnloaded();
     }
 
@@ -318,6 +355,18 @@ export class View extends ViewCommon {
         }
 
         return false;
+    }
+
+    public handleGestureTouch(event: android.view.MotionEvent): any {
+        for (let type in this._gestureObservers) {
+            let list = this._gestureObservers[type];
+            list.forEach(element => {
+                element.androidOnTouchEvent(event);
+            });
+        }
+        if (this.parent instanceof View) {
+            this.parent.handleGestureTouch(event);
+        }
     }
 
     private hasGestureObservers() {
@@ -343,15 +392,20 @@ export class View extends ViewCommon {
     }
 
     private setOnTouchListener() {
-        if (this.nativeViewProtected && this.hasGestureObservers()) {
-            this.touchListenerIsSet = true;
-            if (this.nativeViewProtected.setClickable) {
-                this.nativeViewProtected.setClickable(true);
-            }
+        if (!this.nativeViewProtected || !this.hasGestureObservers()) {
+            return;
+        }
+        
+        // do not set noop listener that handles the event (disabled listener) if IsUserInteractionEnabled is
+        // false as we might need the ability for the event to pass through to a parent view
+        initializeTouchListener();
+        this.touchListener = this.touchListener || new TouchListener(this);
+        this.nativeViewProtected.setOnTouchListener(this.touchListener);
 
-            initializeTouchListener();
-            this.touchListener = this.touchListener || new TouchListener(this);
-            this.nativeViewProtected.setOnTouchListener(this.touchListener);
+        this.touchListenerIsSet = true;
+
+        if (this.nativeViewProtected.setClickable) {
+            this.nativeViewProtected.setClickable(this.isUserInteractionEnabled);
         }
     }
 
@@ -383,6 +437,10 @@ export class View extends ViewCommon {
             return !this.nativeViewProtected.isLayoutRequested();
         }
 
+        return false;
+    }
+
+    get _hasFragments(): boolean {
         return false;
     }
 
@@ -490,7 +548,7 @@ export class View extends ViewCommon {
 
     public getLocationRelativeTo(otherView: ViewCommon): Point {
         if (!this.nativeViewProtected || !this.nativeViewProtected.getWindowToken() ||
-            !otherView.nativeViewProtected || !otherView.nativeViewProtected.getWindowToken() ||
+            !otherView || !otherView.nativeViewProtected || !otherView.nativeViewProtected.getWindowToken() ||
             this.nativeViewProtected.getWindowToken() !== otherView.nativeViewProtected.getWindowToken()) {
             return undefined;
         }
@@ -552,7 +610,7 @@ export class View extends ViewCommon {
         this._dialogFragment = df;
         this._raiseShowingModallyEvent();
 
-        this._dialogFragment.show(parent._getFragmentManager(), this._domId.toString());
+        this._dialogFragment.show(parent._getRootFragmentManager(), this._domId.toString());
     }
 
     protected _hideNativeModalView(parent: View) {
@@ -591,15 +649,8 @@ export class View extends ViewCommon {
     }
 
     [isUserInteractionEnabledProperty.setNative](value: boolean) {
-        if (!value) {
-            initializeDisabledListener();
-            // User interaction is disabled -- we stop it and we do not care whether someone wants to listen for gestures.
-            this.nativeViewProtected.setOnTouchListener(disableUserInteractionListener);
-        } else {
-            this.setOnTouchListener();
-            if (!this.touchListenerIsSet) {
-                this.nativeViewProtected.setOnTouchListener(null);
-            }
+        if (this.nativeViewProtected.setClickable) {
+            this.nativeViewProtected.setClickable(value);
         }
     }
 
@@ -785,7 +836,7 @@ export class View extends ViewCommon {
             androidBackground.onBackgroundOrBorderPropertyChanged(this);
         } else {
             const nativeView = this.nativeViewProtected;
-            org.nativescript.widgets.ViewHelper.setBackground(nativeView, value);
+            nativeView.setBackground(value);
 
             const style = this.style;
             const paddingTop = paddingTopProperty.isSet(style) ? this.effectivePaddingTop : this._defaultPaddingTop;
@@ -804,7 +855,11 @@ export class View extends ViewCommon {
     }
 }
 
-export class CustomLayoutView extends View implements CustomLayoutViewDefinition {
+export class ContainerView extends View {
+    public iosOverflowSafeArea: boolean;
+}
+
+export class CustomLayoutView extends ContainerView implements CustomLayoutViewDefinition {
     nativeViewProtected: android.view.ViewGroup;
 
     public createNativeView() {