+++ tns-core-modules/ui/dialogs/dialogs-common.ts
@@ -2,6 +2,7 @@
 import { View } from "../core/view";
 import { Color } from "../../color";
 import { Page } from "../page";
+import { isIOS } from "../../platform";
 import * as frameModule from "../frame";
 
 export const STRING = "string";
@@ -30,6 +31,31 @@ export module inputType {
      * Email input type.
      */
     export const email: string = "email";
+}
+
+/**
+ * Defines the capitalization type for prompt dialog.
+ */
+export module capitalizationType {
+    /**
+     * No automatic capitalization.
+     */
+    export const none: string = "none";
+
+    /**
+     * Capitalizes every character.
+     */
+    export const all: string = "all";
+
+    /**
+     * Capitalize the first word of each sentence.
+     */
+    export const sentences: string = "sentences";
+
+    /**
+     * Capitalize the first letter of every word.
+     */
+    export const words: string = "words";
 }
 
 let frame: typeof frameModule;
@@ -67,6 +93,9 @@ export function getButtonColors(): { color: Color, backgroundColor: Color } {
     if (!button) {
         const Button = require("ui/button").Button;
         button = new Button;
+        if (isIOS) {
+            button._setupUI({});
+        }
     }
 
     let buttonColor: Color;
@@ -82,6 +111,9 @@ export function getLabelColor(): Color {
     if (!label) {
         const Label = require("ui/label").Label;
         label = new Label;
+        if (isIOS) {
+            label._setupUI({});
+        }
     }
 
     let labelColor: Color;
@@ -95,6 +127,9 @@ export function getTextFieldColor(): Color {
     if (!textField) {
         const TextField = require("ui/text-field").TextField;
         textField = new TextField();
+        if (isIOS) {
+            textField._setupUI({});
+        }
     }
 
     let textFieldColor: Color;