+++ tns-core-modules/ui/dialogs/dialogs.ios.ts
@@ -3,7 +3,7 @@
  */
 import { View, ios as iosView } from "../core/view";
 import { ConfirmOptions, PromptOptions, PromptResult, LoginOptions, LoginResult, ActionOptions } from ".";
-import { getCurrentPage, getLabelColor, getButtonColors, getTextFieldColor, isDialogOptions, inputType, ALERT, OK, CONFIRM, CANCEL, PROMPT, LOGIN } from "./dialogs-common";
+import { getCurrentPage, getLabelColor, getButtonColors, getTextFieldColor, isDialogOptions, inputType, capitalizationType, ALERT, OK, CONFIRM, CANCEL, PROMPT, LOGIN } from "./dialogs-common";
 import { isString, isDefined, isFunction } from "../../utils/types";
 import { getRootView } from "../../application";
 
@@ -113,6 +113,23 @@ export function prompt(arg: any): Promise<PromptResult> {
             });
 
             textField = alertController.textFields.firstObject;
+
+            if (options) {
+                switch (options.capitalizationType) {
+                    case capitalizationType.all: {
+                        textField.autocapitalizationType = UITextAutocapitalizationType.AllCharacters; break;
+                    }
+                    case capitalizationType.sentences: {
+                        textField.autocapitalizationType = UITextAutocapitalizationType.Sentences; break;
+                    }
+                    case capitalizationType.words: {
+                        textField.autocapitalizationType = UITextAutocapitalizationType.Words; break;
+                    }
+                    default: {
+                        textField.autocapitalizationType = UITextAutocapitalizationType.None;
+                    }
+                }
+            }
 
             addButtonsToAlertController(alertController, options,
                 (r) => { resolve({ result: r, text: textField.text }); });