+++ tns-core-modules/ui/tab-view/tab-view.ios.ts
@@ -5,14 +5,18 @@ import { ios as iosView, ViewBase } from "../core/view";
 import {
     TabViewBase, TabViewItemBase, itemsProperty, selectedIndexProperty,
     tabTextColorProperty, tabTextFontSizeProperty, tabBackgroundColorProperty, selectedTabTextColorProperty, iosIconRenderingModeProperty,
-    View, fontInternalProperty, layout, traceEnabled, traceWrite, traceCategories, Color
+    View, fontInternalProperty, layout, traceEnabled, traceWrite, traceCategories, Color, traceMissingIcon
 } from "./tab-view-common"
 import { textTransformProperty, TextTransform, getTransformedText } from "../text-base";
 import { fromFileOrResource } from "../../image-source";
 import { profile } from "../../profiling";
 import { Frame } from "../frame";
-
+import { ios as iosUtils } from "../../utils/utils"
+import { device } from "../../platform";
 export * from "./tab-view-common";
+
+const majorVersion = iosUtils.MajorVersion;
+const isPhone = device.deviceType === "Phone";
 
 class UITabBarControllerImpl extends UITabBarController {
 
@@ -46,6 +50,17 @@ class UITabBarControllerImpl extends UITabBarController {
         if (owner && !owner.parent && owner.isLoaded && !this.presentedViewController) {
             owner.callUnloaded();
         }
+    }
+
+    public viewWillTransitionToSizeWithTransitionCoordinator(size: CGSize, coordinator: UIViewControllerTransitionCoordinator): void {
+        super.viewWillTransitionToSizeWithTransitionCoordinator(size, coordinator);
+        UIViewControllerTransitionCoordinator.prototype.animateAlongsideTransitionCompletion
+            .call(coordinator, null, () => {
+                const owner = this._owner.get();
+                if (owner && owner.items) {
+                    owner.items.forEach(tabItem => tabItem._updateTitleAndIconPositions());
+                }
+            });
     }
 }
 
@@ -129,16 +144,32 @@ class UINavigationControllerDelegateImpl extends NSObject implements UINavigatio
     }
 }
 
-function updateItemTitlePosition(tabBarItem: UITabBarItem): void {
-    if (typeof (<any>tabBarItem).setTitlePositionAdjustment === "function") {
-        (<any>tabBarItem).setTitlePositionAdjustment({ horizontal: 0, vertical: -20 });
-    } else {
-        tabBarItem.titlePositionAdjustment = { horizontal: 0, vertical: -20 };
+function updateTitleAndIconPositions(tabItem: TabViewItem, tabBarItem: UITabBarItem, controller: UIViewController) {
+    if (!tabItem || !tabBarItem) {
+        return;
     }
-}
 
-function updateItemIconPosition(tabBarItem: UITabBarItem): void {
-    tabBarItem.imageInsets = new UIEdgeInsets({ top: 6, left: 0, bottom: -6, right: 0 });
+    // For iOS <11 icon is *always* above the text.
+    // For iOS 11 icon is above the text *only* on phones in portrait mode.
+    const orientation = controller.interfaceOrientation;
+    const isPortrait = orientation !== UIInterfaceOrientation.LandscapeLeft && orientation !== UIInterfaceOrientation.LandscapeRight;
+    const isIconAboveTitle = (majorVersion < 11) || (isPhone && isPortrait);
+
+    if (!tabItem.iconSource) {
+        if (isIconAboveTitle) {
+            tabBarItem.titlePositionAdjustment = { horizontal: 0, vertical: -20 };
+        } else {
+            tabBarItem.titlePositionAdjustment = { horizontal: 0, vertical: 0 };
+        }
+    }
+
+    if (!tabItem.title) {
+        if (isIconAboveTitle) {
+            tabBarItem.imageInsets = new UIEdgeInsets({ top: 6, left: 0, bottom: -6, right: 0 });
+        } else {
+            tabBarItem.imageInsets = new UIEdgeInsets({ top: 0, left: 0, bottom: 0, right: 0 });
+        }
+    }
 }
 
 export class TabViewItem extends TabViewItemBase {
@@ -174,11 +205,7 @@ export class TabViewItem extends TabViewItemBase {
             const title = getTransformedText(this.title, this.style.textTransform);
 
             const tabBarItem = UITabBarItem.alloc().initWithTitleImageTag(title, icon, index);
-            if (!icon) {
-                updateItemTitlePosition(tabBarItem);
-            } else if (!title) {
-                updateItemIconPosition(tabBarItem);
-            }
+            updateTitleAndIconPositions(this, tabBarItem, controller);
 
             // TODO: Repeating code. Make TabViewItemBase - ViewBase and move the colorProperty on tabViewItem.
             // Delete the repeating code.
@@ -188,6 +215,13 @@ export class TabViewItem extends TabViewItemBase {
         }
     }
 
+    public _updateTitleAndIconPositions() {
+        if (!this.__controller || !this.__controller.tabBarItem) {
+            return;
+        }
+        updateTitleAndIconPositions(this, this.__controller.tabBarItem, this.__controller);
+    }
+
     [textTransformProperty.setNative](value: TextTransform) {
         this._update();
     }
@@ -195,6 +229,7 @@ export class TabViewItem extends TabViewItemBase {
 
 export class TabView extends TabViewBase {
     public viewController: UITabBarControllerImpl;
+    public items: TabViewItem[];
     public _ios: UITabBarControllerImpl;
     private _delegate: UITabBarControllerDelegateImpl;
     private _moreNavigationControllerDelegate: UINavigationControllerDelegateImpl;
@@ -205,9 +240,18 @@ export class TabView extends TabViewBase {
 
         this.viewController = this._ios = UITabBarControllerImpl.initWithOwner(new WeakRef(this));
         this.nativeViewProtected = this._ios.view;
+    }
+
+    initNativeView() {
+        super.initNativeView();
         this._delegate = UITabBarControllerDelegateImpl.initWithOwner(new WeakRef(this));
         this._moreNavigationControllerDelegate = UINavigationControllerDelegateImpl.initWithOwner(new WeakRef(this));
-        //This delegate is set on the last line of _addTabs method.
+    }
+
+    disposeNativeView() {
+        this._delegate = null;
+        this._moreNavigationControllerDelegate = null;
+        super.disposeNativeView();
     }
 
     @profile
@@ -367,11 +411,7 @@ export class TabView extends TabViewBase {
             const controller = this.getViewController(item);
             const icon = this._getIcon(item.iconSource);
             const tabBarItem = UITabBarItem.alloc().initWithTitleImageTag((item.title || ""), icon, i);
-            if (!icon) {
-                updateItemTitlePosition(tabBarItem);
-            } else if (!item.title) {
-                updateItemIconPosition(tabBarItem);
-            }
+            updateTitleAndIconPositions(item, tabBarItem, controller);
 
             applyStatesToItem(tabBarItem, states);
 
@@ -411,6 +451,8 @@ export class TabView extends TabViewBase {
                 const originalRenderedImage = is.ios.imageWithRenderingMode(this._getIconRenderingMode());
                 this._iconsCache[iconSource] = originalRenderedImage;
                 image = originalRenderedImage;
+            } else {
+                traceMissingIcon(iconSource);
             }
         }