+++ tns-core-modules/ui/text-base/text-base.ios.ts
@@ -11,6 +11,7 @@ export * from "./text-base-common";
 export class TextBase extends TextBaseCommon {
 
     public nativeViewProtected: UITextField | UITextView | UILabel | UIButton;
+    public nativeTextViewProtected: UITextField | UITextView | UILabel | UIButton;
 
     [textProperty.getDefault](): number | symbol {
         return resetSymbol;
@@ -32,7 +33,7 @@ export class TextBase extends TextBaseCommon {
     }
 
     [colorProperty.getDefault](): UIColor {
-        let nativeView = this.nativeViewProtected;
+        let nativeView = this.nativeTextViewProtected;
         if (nativeView instanceof UIButton) {
             return nativeView.titleColorForState(UIControlState.Normal);
         } else {
@@ -41,7 +42,7 @@ export class TextBase extends TextBaseCommon {
     }
     [colorProperty.setNative](value: Color | UIColor) {
         const color = value instanceof Color ? value.ios : value;
-        const nativeView = this.nativeViewProtected;
+        const nativeView = this.nativeTextViewProtected;
         if (nativeView instanceof UIButton) {
             nativeView.setTitleColorForState(color, UIControlState.Normal);
             nativeView.titleLabel.textColor = color;
@@ -51,13 +52,13 @@ export class TextBase extends TextBaseCommon {
     }
 
     [fontInternalProperty.getDefault](): UIFont {
-        let nativeView = this.nativeViewProtected;
+        let nativeView = this.nativeTextViewProtected;
         nativeView = nativeView instanceof UIButton ? nativeView.titleLabel : nativeView;
         return nativeView.font;
     }
     [fontInternalProperty.setNative](value: Font | UIFont) {
         if (!(value instanceof Font) || !this.formattedText) {
-            let nativeView = this.nativeViewProtected;
+            let nativeView = this.nativeTextViewProtected;
             nativeView = nativeView instanceof UIButton ? nativeView.titleLabel : nativeView;
             const font = value instanceof Font ? value.getUIFont(nativeView.font) : value;
             nativeView.font = font;
@@ -65,7 +66,7 @@ export class TextBase extends TextBaseCommon {
     }
 
     [textAlignmentProperty.setNative](value: TextAlignment) {
-        const nativeView = <UITextField | UITextView | UILabel>this.nativeViewProtected;
+        const nativeView = <UITextField | UITextView | UILabel>this.nativeTextViewProtected;
         switch (value) {
             case "initial":
             case "left":
@@ -98,7 +99,7 @@ export class TextBase extends TextBaseCommon {
 
     _setNativeText(reset: boolean = false): void {
         if (reset) {
-            const nativeView = this.nativeViewProtected;
+            const nativeView = this.nativeTextViewProtected;
             if (nativeView instanceof UIButton) {
                 // Clear attributedText or title won't be affected.
                 nativeView.setAttributedTitleForState(null, UIControlState.Normal);
@@ -123,26 +124,26 @@ export class TextBase extends TextBaseCommon {
         const attrText = this.createNSMutableAttributedString(this.formattedText);
         // TODO: letterSpacing should be applied per Span.
         if (this.letterSpacing !== 0) {
-            attrText.addAttributeValueRange(NSKernAttributeName, this.letterSpacing * this.nativeViewProtected.font.pointSize, { location: 0, length: attrText.length });
+            attrText.addAttributeValueRange(NSKernAttributeName, this.letterSpacing * this.nativeTextViewProtected.font.pointSize, { location: 0, length: attrText.length });
         }
 
         if (this.style.lineHeight) {
             const paragraphStyle = NSMutableParagraphStyle.alloc().init();
             paragraphStyle.lineSpacing = this.lineHeight;
             // make sure a possible previously set text alignment setting is not lost when line height is specified
-            paragraphStyle.alignment = (<UITextField | UITextView | UILabel>this.nativeViewProtected).textAlignment;
-            if (this.nativeViewProtected instanceof  UILabel) {
+            paragraphStyle.alignment = (<UITextField | UITextView | UILabel>this.nativeTextViewProtected).textAlignment;
+            if (this.nativeTextViewProtected instanceof  UILabel) {
                 // make sure a possible previously set line break mode is not lost when line height is specified
-                paragraphStyle.lineBreakMode = this.nativeViewProtected.lineBreakMode;
+                paragraphStyle.lineBreakMode = this.nativeTextViewProtected.lineBreakMode;
             }
             attrText.addAttributeValueRange(NSParagraphStyleAttributeName, paragraphStyle, { location: 0, length: attrText.length });
         }
 
-        if (this.nativeViewProtected instanceof UIButton) {
-            this.nativeViewProtected.setAttributedTitleForState(attrText, UIControlState.Normal);
+        if (this.nativeTextViewProtected instanceof UIButton) {
+            this.nativeTextViewProtected.setAttributedTitleForState(attrText, UIControlState.Normal);
         }
         else {
-            this.nativeViewProtected.attributedText = attrText;
+            this.nativeTextViewProtected.attributedText = attrText;
         }
     }
 
@@ -167,22 +168,22 @@ export class TextBase extends TextBaseCommon {
         }
 
         if (style.letterSpacing !== 0) {
-            dict.set(NSKernAttributeName, style.letterSpacing * this.nativeViewProtected.font.pointSize);
+            dict.set(NSKernAttributeName, style.letterSpacing * this.nativeTextViewProtected.font.pointSize);
         }
 
         if (style.lineHeight) {
             const paragraphStyle = NSMutableParagraphStyle.alloc().init();
             paragraphStyle.lineSpacing = style.lineHeight;
             // make sure a possible previously set text alignment setting is not lost when line height is specified
-            paragraphStyle.alignment = (<UITextField | UITextView | UILabel>this.nativeViewProtected).textAlignment;
-            if (this.nativeViewProtected instanceof  UILabel) {
+            paragraphStyle.alignment = (<UITextField | UITextView | UILabel>this.nativeTextViewProtected).textAlignment;
+            if (this.nativeTextViewProtected instanceof  UILabel) {
                 // make sure a possible previously set line break mode is not lost when line height is specified
-                paragraphStyle.lineBreakMode = this.nativeViewProtected.lineBreakMode;
+                paragraphStyle.lineBreakMode = this.nativeTextViewProtected.lineBreakMode;
             }
             dict.set(NSParagraphStyleAttributeName, paragraphStyle);
         }
 
-        const isTextView = this.nativeViewProtected instanceof UITextView;
+        const isTextView = this.nativeTextViewProtected instanceof UITextView;
         if (style.color && (dict.size > 0 || isTextView)) {
             dict.set(NSForegroundColorAttributeName, style.color.ios);
         }
@@ -193,25 +194,25 @@ export class TextBase extends TextBaseCommon {
         if (dict.size > 0 || isTextView) {
             if (isTextView) {
                 // UITextView's font seems to change inside.
-                dict.set(NSFontAttributeName, this.nativeViewProtected.font);
+                dict.set(NSFontAttributeName, this.nativeTextViewProtected.font);
             }
             
             const result = NSMutableAttributedString.alloc().initWithString(source);
             result.setAttributesRange(<any>dict, { location: 0, length: source.length });
-            if (this.nativeViewProtected instanceof UIButton) {
-                this.nativeViewProtected.setAttributedTitleForState(result, UIControlState.Normal);
+            if (this.nativeTextViewProtected instanceof UIButton) {
+                this.nativeTextViewProtected.setAttributedTitleForState(result, UIControlState.Normal);
             } else {
-                this.nativeViewProtected.attributedText = result;
+                this.nativeTextViewProtected.attributedText = result;
             }
         } else {
-            if (this.nativeViewProtected instanceof UIButton) {
+            if (this.nativeTextViewProtected instanceof UIButton) {
                 // Clear attributedText or title won't be affected.
-                this.nativeViewProtected.setAttributedTitleForState(null, UIControlState.Normal);
-                this.nativeViewProtected.setTitleForState(source, UIControlState.Normal);
+                this.nativeTextViewProtected.setAttributedTitleForState(null, UIControlState.Normal);
+                this.nativeTextViewProtected.setTitleForState(source, UIControlState.Normal);
             } else {
                 // Clear attributedText or text won't be affected.
-                this.nativeViewProtected.attributedText = undefined;
-                this.nativeViewProtected.text = source;
+                this.nativeTextViewProtected.attributedText = undefined;
+                this.nativeTextViewProtected.text = source;
             }
         }
     }
@@ -237,7 +238,7 @@ export class TextBase extends TextBaseCommon {
     }
 
     createMutableStringForSpan(span: Span, text: string): NSMutableAttributedString {
-        const viewFont = this.nativeViewProtected.font;
+        const viewFont = this.nativeTextViewProtected.font;
         let attrDict = <{ key: string, value: any }>{};
         const style = span.style;
         const bold = isBold(style.fontWeight);